pyproject.toml
[project]
name = "enterprise-data-product-planner"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "pydantic>=2.0",
    "sqlglot>=23.0",
    "neo4j>=5.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0.3",
    "anthropic>=0.40",
    "streamlit>=1.35",
    "plotly>=5.20",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "hypothesis>=6.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "graph", "ontology", "storage", "api", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = false

[dependency-groups]
dev = [
    "hypothesis>=6.151.14",
]
